requirements.txt
fastapi
uvicorn
pinecone
onnxruntime-gpu
numpy
openai
pyyaml
pydantic_settings
transformers
pinecone_text